// src/utils/types.ts
type NepaliDate = {
    year: number | null;
    month: number | null;
    date: number | null;
    day: string | null;
    nmonth: string | null;
    num_day: number | null;
}

type EnglishDate = {
    year: number | null;
    month: number | null;
    date: number | null;
    day: string | null;
    nmonth: string | null;
    num_day: number | null;
}

export type {
    NepaliDate,
    EnglishDate
}